JanDrishti-Pollution-Action-DashBoard/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
supabase==2.10.0
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic[email]==2.9.2
email-validator==2.1.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
httpx==0.27.2
groq==0.4.1
pandas
geopandas
shapely
requests
folium
rtree
pyproj
redis==5.0.1
apscheduler==3.10.4
twilio==9.3.0

